Creating SAFE extractor - an object used for surrogate-assisted feature extraction

The safe_extraction() function creates a SAFE-extractor object which may be used later for surrogate feature extraction.

safe_extraction(explainer, response_type = "ale", penalty = "MBIC",
  no_segments = 2, interactions = TRUE, inter_param = 2,
  inter_threshold = 0.4, verbose = TRUE)

Arguments

explainer
DALEX explainer created with explain() function

response_type
character, type of response to be calculated, one of: "pdp", "ale". If features are uncorrelated, one can use "pdp" type - otherwise "ale" is strongly recommended.

penalty
penalty for introducing another changepoint, one of "AIC", "BIC", "SIC", "MBIC", "Hannan-Quinn" or numeric non-negative value

no_segments
numeric, a number of segments variable is to be divided into in case of founding no breakpoints

interactions
logical, if interactions between variables are to be taken into account

inter_param
numeric, a positive value indicating which of single observation non-additive effects are to be regarded as significant, the higher value the higher non-additive effect has to be to be taken into account

inter_threshold
numeric, a value from [0,1] interval indicating which interactions should be returned as significant. It corresponds to the percentage of observations for which interaction measure is greater than inter_param - if this percentage is less than inter_threshold then interaction effect is ignored.

verbose
logical, if progress bar is to be printed

Value

safe_extractor object containing information about variables transformation

See also