Identyfying changes using a nonparametric cost function

The safely_detect_changepoints() function calculates the optimal positioning and number of changepoints for given data and penalty. It uses a PELT algorithm with a nonparametric cost function based on the empirical distribution.

safely_detect_changepoints(data, penalty = "MBIC", nquantiles = 10)

Arguments

data
a vector within which you wish to find changepoints

penalty
penalty for introducing another changepoint, one of "AIC", "BIC", "SIC", "MBIC", "Hannan-Quinn" or numeric non-negative value

nquantiles
the number of quantiles used in integral approximation

Value

a vector of optimal changepoint positions (last observations of each segment)

See also