Feature selection on the dataset with transformed variables

The safely_select_variables() function selects variables from dataset returned by safely_transform_data() function. For each original variable:

if encoding = "categorical" then exactly one variable is chosen

either original one or transformed one. The choice is based on the AIC value for linear model (regression) or logistic regression (classification).

if encoding = "one-hot" then backward feature selection is performed on all features (using AIC value) and the best set of variables is chosen.

safely_select_variables(safe_extractor, data, y = NULL, which_y = NULL,
  encoding = "categorical", verbose = TRUE, class_pred = NULL)

Arguments

safe_extractor
object containing information about variables transformations created with safe_extraction() function

data
data, original dataset or the one returned by safely_transform_data() function. If data do not contain transformed variables then transformation is done inside this function using 'safe_extractor' argument. Data may contain response variable or not - if it does then 'which_y' argument must be given, otherwise 'y' argument should be provided.

y
vector of responses, must be given if data does not contain it

which_y
numeric or character (optional), must be given if data contains response values

encoding
method of representing factor variables in dataset, one of: "categorical", "one-hot"

verbose
logical, if progress bar is to be printed

class_pred
numeric or character, used only in multi-classification problems. If response vector has more than two levels, then 'class_pred' should indicate the class of interest which will denote failure - all other classess will stand for success.

Value

vector of variables names, selected based on AIC values